+++ Sources/Matrix_Mx1.swift
@@ -1,1 +1,1 @@
-public typealias ColumnVector<M: Size, Real: La.Real> = Matrix<M, _1, Real>
+public typealias ColumnVector<M: Size, Real> = Matrix<M, _1, Real>